refactor: autoDetectLabel fn to helper file; tests passing
