+++ packages/auto-label/src/helper.ts
@@ -16,7 +16,64 @@
 // eslint-disable-next-line node/no-extraneous-import
 import {Context} from 'probot';
 import {logger} from 'gcf-utils';
-import {LanguageConfig, PathConfig} from './auto-label';
+import {DriftApi, LanguageConfig, PathConfig} from './auto-label';
+
+// Helper functions for product based labels
+
+// autoDetectLabel tries to detect the right api: label based on the issue
+// title.
+//
+// For example, an issue titled `spanner/transactions: TestSample failed` would
+// be labeled `api: spanner`.
+export function autoDetectLabel(
+    apis: DriftApi[] | null,
+    title: string
+): string | undefined {
+  if (!apis || !title) {
+    return undefined;
+  }
+
+  // The Conventional Commits "docs:" and "build:" prefixes are far more common
+  // than the APIs. So, never label those with "api: docs" or "api: build".
+  if (title.startsWith('docs:') || title.startsWith('build:')) {
+    return undefined;
+  }
+
+  // Regex to match the scope of a Conventional Commit message.
+  const conv = /[^(]+\(([^)]+)\):/;
+  const match = title.match(conv);
+
+  let firstPart = match ? match[1] : title;
+
+  // Remove common prefixes. For example,
+  // https://github.com/GoogleCloudPlatform/java-docs-samples/issues/3578.
+  const trimPrefixes = ['com.example.', 'com.google.', 'snippets.'];
+  for (const prefix of trimPrefixes) {
+    if (firstPart.startsWith(prefix)) {
+      firstPart = firstPart.slice(prefix.length);
+    }
+  }
+
+  if (firstPart.startsWith('/')) firstPart = firstPart.substr(1); // Remove leading /.
+  firstPart = firstPart.split(':')[0]; // Before the colon, if there is one.
+  firstPart = firstPart.split('/')[0]; // Before the slash, if there is one.
+  firstPart = firstPart.split('.')[0]; // Before the period, if there is one.
+  firstPart = firstPart.split('_')[0]; // Before the underscore, if there is one.
+  firstPart = firstPart.toLowerCase(); // Convert to lower case.
+  firstPart = firstPart.replace(/\s/, ''); // Remove spaces.
+
+  // Replace some known firstPart values with their API name.
+  const commonConversions = new Map();
+  commonConversions.set('video', 'videointelligence');
+  firstPart = commonConversions.get(firstPart) || firstPart;
+
+  // Some APIs have "cloud" before the name (e.g. cloudkms and cloudiot).
+  const possibleLabels = [`api: ${firstPart}`, `api: cloud${firstPart}`];
+  return apis.find(api => possibleLabels.indexOf(api.github_label) > -1)
+      ?.github_label;
+}
+
+// Helper functions for language and path based labels
 
 // A mapping of languages to their file extensions
 import defaultExtensions from './extensions.json';
